Task 2 - Using Puppet, create a manifest that kills a process named killmenow

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -4,4 +4,6 @@
 exec { 'kill_killmenow':
   command  => 'pkill killmenow',    #executes the kill command pkill on killmenow
   path     => ['/bin', '/usr/bin'], # Specify the command's path as needed
+  refreshonly => true,
+  onlyif      => 'pgrep killmenow',    # Check if the process exists before attempting to kill it
 }
